# frontend/openapi.json
{
  "components": {
    "schemas": {
      "Aircraft": {
        "description": "One aircraft state vector, normalised across providers.\n\nUnits are SI throughout (metres, metres per second). Conversion to feet or\nknots is a presentation concern and belongs to the client.",
        "properties": {
          "altitude_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Geometric, else barometric",
            "title": "Altitude M"
          },
          "callsign": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Callsign"
          },
          "heading_deg": {
            "anyOf": [
              {
                "exclusiveMaximum": 360.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading Deg"
          },
          "icao24": {
            "description": "ICAO 24-bit address, hex",
            "maxLength": 6,
            "minLength": 6,
            "title": "Icao24",
            "type": "string"
          },
          "last_contact": {
            "format": "date-time",
            "title": "Last Contact",
            "type": "string"
          },
          "on_ground": {
            "default": false,
            "title": "On Ground",
            "type": "boolean"
          },
          "origin_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Country"
          },
          "position": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "registration": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration"
          },
          "source": {
            "description": "Provider that produced this state",
            "title": "Source",
            "type": "string"
          },
          "squawk": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Squawk"
          },
          "type_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ICAO aircraft type designator",
            "title": "Type Code"
          },
          "velocity_ms": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Velocity Ms"
          },
          "vertical_rate_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vertical Rate Ms"
          }
        },
        "required": [
          "icao24",
          "callsign",
          "registration",
          "type_code",
          "squawk",
          "origin_country",
          "position",
          "altitude_m",
          "velocity_ms",
          "heading_deg",
          "vertical_rate_ms",
          "on_ground",
          "last_contact",
          "source"
        ],
        "title": "Aircraft",
        "type": "object"
      },
      "AircraftCollection": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Aircraft"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "count",
          "items"
        ],
        "title": "AircraftCollection",
        "type": "object"
      },
      "AircraftTrack": {
        "properties": {
          "callsign": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Callsign"
          },
          "icao24": {
            "title": "Icao24",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TrackPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "icao24",
          "callsign",
          "points",
          "source"
        ],
        "title": "AircraftTrack",
        "type": "object"
      },
...
      "AssetOut": {
        "properties": {
          "candles": {
            "items": {
              "$ref": "#/components/schemas/Candle"
            },
            "title": "Candles",
            "type": "array"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "delayed": {
            "default": true,
            "title": "Delayed",
            "type": "boolean"
          },
          "exchange": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange"
          },
          "instrument": {
            "$ref": "#/components/schemas/Instrument"
          },
          "range": {
            "$ref": "#/components/schemas/HistoryRange"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "technicals": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Technicals"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "instrument",
          "currency",
          "exchange",
          "range",
          "candles",
          "technicals",
          "source"
        ],
        "title": "AssetOut",
        "type": "object"
      },
...
      "CalendarEvent": {
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "currency": {
            "description": "Currency the release concerns (FX-desk convention)",
            "title": "Currency",
            "type": "string"
          },
          "forecast": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast"
          },
          "impact": {
            "$ref": "#/components/schemas/Impact"
          },
          "previous": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "currency",
          "at",
          "impact",
          "forecast",
          "previous"
        ],
        "title": "CalendarEvent",
        "type": "object"
      },
...
      "CryptoAsset": {
        "properties": {
          "change_24h_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change 24H Pct"
          },
          "change_7d_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change 7D Pct"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "market_cap_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Usd"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price_usd": {
            "title": "Price Usd",
            "type": "number"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "volume_24h_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume 24H Usd"
          }
        },
        "required": [
          "id",
          "symbol",
          "name",
          "price_usd",
          "change_24h_pct",
          "change_7d_pct",
          "market_cap_usd",
          "volume_24h_usd",
          "updated_at",
          "source"
        ],
        "title": "CryptoAsset",
        "type": "object"
      },
...
      "FundingBoard": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/FundingRate"
            },
            "title": "Items",
            "type": "array"
          },
          "missing": {
            "items": {
              "$ref": "#/components/schemas/MissingOut"
            },
            "title": "Missing",
            "type": "array"
          }
        },
        "required": [
          "count",
          "items",
          "missing"
        ],
        "title": "FundingBoard",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HistoryRange": {
        "enum": [
          "1mo",
          "3mo",
          "6mo",
          "1y",
          "2y",
          "5y"
        ],
        "title": "HistoryRange",
        "type": "string"
      },
...
      "LiquidationBoard": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/LiquidationSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "missing": {
            "items": {
              "$ref": "#/components/schemas/MissingOut"
            },
            "title": "Missing",
            "type": "array"
          }
        },
        "required": [
          "count",
          "items",
          "missing"
        ],
        "title": "LiquidationBoard",
        "type": "object"
      },
...
      "PredictionMarket": {
        "properties": {
          "end_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "event_title": {
            "title": "Event Title",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "one_day_change": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Change of the first outcome",
            "title": "One Day Change"
          },
          "outcomes": {
            "items": {
              "$ref": "#/components/schemas/Outcome"
            },
            "title": "Outcomes",
            "type": "array"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "volume_24h_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume 24H Usd"
          },
          "volume_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume Usd"
          }
        },
        "required": [
          "id",
          "question",
          "event_title",
          "outcomes",
          "one_day_change",
          "volume_usd",
          "volume_24h_usd",
          "end_date",
          "url",
          "source"
        ],
        "title": "PredictionMarket",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Argus API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/aviation/aircraft": {
      "get": {
        "operationId": "list_aircraft_api_v1_aviation_aircraft_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_on_ground",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include On Ground",
              "type": "boolean"
            }
          },
          {
            "description": "west,south,east,north",
            "in": "query",
            "name": "bbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "west,south,east,north",
              "examples": [
                "2.0,48.5,2.8,49.1"
              ],
              "title": "Bbox"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AircraftCollection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Aircraft",
        "tags": [
          "aviation"
        ]
      }
    },
    "/api/v1/aviation/aircraft/{icao24}/track": {
      "get": {
        "description": "Today's positions for one aircraft, oldest first (at most 2,000 points).",
        "operationId": "aircraft_track_api_v1_aviation_aircraft__icao24__track_get",
        "parameters": [
          {
            "description": "ICAO 24-bit address",
            "in": "path",
            "name": "icao24",
            "required": true,
            "schema": {
              "description": "ICAO 24-bit address",
              "pattern": "^[0-9a-fA-F]{6}$",
              "title": "Icao24",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AircraftTrack"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Aircraft Track",
        "tags": [
          "aviation"
        ]
      }
    },
    "/api/v1/aviation/military": {
      "get": {
        "description": "Aircraft flagged military by the aggregator's database. Worldwide unless `bbox`.",
        "operationId": "list_military_api_v1_aviation_military_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_on_ground",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include On Ground",
              "type": "boolean"
            }
          },
          {
            "description": "west,south,east,north",
            "in": "query",
            "name": "bbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "west,south,east,north",
              "examples": [
                "2.0,48.5,2.8,49.1"
              ],
              "title": "Bbox"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AircraftCollection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Military",
        "tags": [
          "aviation"
        ]
      }
    },
...
    "/api/v1/macro/calendar": {
      "get": {
        "operationId": "calendar_api_v1_macro_calendar_get",
        "parameters": [
          {
            "in": "query",
            "name": "impact",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(high|medium|low|holiday)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Impact"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CalendarEvent"
                  },
                  "title": "Response Calendar Api V1 Macro Calendar Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calendar",
        "tags": [
          "finance"
        ]
      }
    },
...
    "/api/v1/markets/assets/{symbol}": {
      "get": {
        "description": "Daily bars and a technical reading (descriptive, not advice) for any Yahoo symbol.",
        "operationId": "asset_api_v1_markets_assets__symbol__get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9^=.\\-]{1,20}$",
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "range",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/HistoryRange",
              "default": "1y"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Asset",
        "tags": [
          "finance"
        ]
      }
    },
    "/api/v1/markets/crypto": {
      "get": {
        "operationId": "crypto_api_v1_markets_crypto_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CryptoAsset"
                  },
                  "title": "Response Crypto Api V1 Markets Crypto Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Crypto",
        "tags": [
          "finance"
        ]
      }
    },
    "/api/v1/markets/funding": {
      "get": {
        "description": "Perpetual funding: positive = longs pay shorts (crowded long).",
        "operationId": "funding_api_v1_markets_funding_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbols",
            "required": false,
            "schema": {
              "default": "BTC,ETH,SOL",
              "pattern": "^[A-Za-z]{2,10}(,[A-Za-z]{2,10}){0,9}$",
              "title": "Symbols",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundingBoard"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Funding",
        "tags": [
          "finance"
        ]
      }
    },
    "/api/v1/markets/liquidations": {
      "get": {
        "description": "Recent forced liquidations on one exchange: a sample, not the whole market.",
        "operationId": "liquidations_api_v1_markets_liquidations_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbols",
            "required": false,
            "schema": {
              "default": "BTC,ETH,SOL",
              "pattern": "^[A-Za-z]{2,10}(,[A-Za-z]{2,10}){0,9}$",
              "title": "Symbols",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiquidationBoard"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Liquidations",
        "tags": [
          "finance"
        ]
      }
    },
    "/api/v1/markets/prediction": {
      "get": {
        "description": "Open prediction markets. Prices are crowd probabilities, not forecasts of record.",
        "operationId": "prediction_api_v1_markets_prediction_get",
        "parameters": [
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "default": "geopolitics",
              "pattern": "^[a-z0-9-]+$",
              "title": "Tag",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PredictionMarket"
                  },
                  "title": "Response Prediction Api V1 Markets Prediction Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prediction",
        "tags": [
          "finance"
        ]
      }
    },
...
  }
}